+++ ci_cd/pipe_embedded/locals.tf
@@ -10,7 +10,7 @@ locals {
   build_name_list_map = {
     for k, v_site in local.lx_map : k => flatten([
       for v_stage in concat(v_site.build_stage_list, v_site.deploy_stage_list) : [
-        for _, v_action in v_stage.action_map : v_action.configuration_project_name
+        for _, v_action in v_stage.action_map : v_action.configuration_build_project_name if v_action.category == "Build"
       ]
     ])
   }
@@ -24,11 +24,38 @@ locals {
       )
     })
   }
+  deploy_group_to_app_map = {
+    for k, v_site in local.lx_map : k => merge([
+      for v_stage in concat(v_site.build_stage_list, v_site.deploy_stage_list) : merge([
+        for _, v_action in v_stage.action_map : {
+          (v_action.configuration_deploy_group_name) = (v_action.configuration_deploy_application_name)
+        } if v_action.category == "Deploy"
+      ]...)
+    ]...)
+  }
   l0_map = {
     for k, v in var.pipe_map : k => v
   }
   l1_map = {
     for k, v in local.l0_map : k => merge(v, module.name_map.data[k], {
+      build_stage_list = [
+        for v_build in v.build_stage_list : merge(v_build, {
+          action_map = {
+            for k_act, v_act in v_build.action_map : k_act => merge(v_act, {
+              category = v_act.category == null ? var.pipe_stage_category_default : v_act.category
+            })
+          }
+        })
+      ]
+      deploy_stage_list = [
+        for v_build in v.deploy_stage_list : merge(v_build, {
+          action_map = {
+            for k_act, v_act in v_build.action_map : k_act => merge(v_act, {
+              category = v_act.category == null ? var.pipe_stage_category_default : v_act.category
+            })
+          }
+        })
+      ]
     })
   }
   l2_map = {