+++ name_map/variables.tf
@@ -2,6 +2,7 @@ variable "name_map" {
   type = map(object({
     name_include_app_fields = optional(bool)
     name_infix              = optional(bool)
+    name_override           = optional(string)
     name_prefix             = optional(string)
     name_suffix             = optional(string)
     tags                    = optional(map(string))